refactor: separate filter logic in utils
