+++ src/pages/AppInventoryList/index.tsx
@@ -13,7 +13,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { setAppInfo, setFilters } from './../../redux/slices/appInfo.slice';
 import AutoCompleteSearch from './../../components/AutoCompleteSearch';
 import { convertAvails } from './../../utils/convertAvails';
-import { useSorts } from './../../utils/sortFilters';
+import { useSorts, useFilters } from './../../utils/sortFilters';
 
 const AppInventoryList = () => {
 
@@ -119,6 +119,7 @@ const AppInventoryList = () => {
     navigate(`/edit-app`)
   }
 
+  // handle pagination, sorts and filters when change
   const handleTableChange = (pagination: any, tableFilters: any, sorter: any, extra: any) => {
     switch (extra['action']) {
       case 'sort':
@@ -131,76 +132,7 @@ const AppInventoryList = () => {
         setRequestBody({ ...requestBody, pageIndex: pagination.current - 1, pageSize: pagination.pageSize });
         break;
       case 'filter':
-        const allFilters: any = Object.keys(tableFilters);
-        const updatedFilters: IFilter[] = [];
-        if (allFilters.length > 0) {
-          allFilters?.forEach((filter: string) => {
-            if(tableFilters[filter]) {
-              let singleFilter = {
-                name: '',
-                value: '' as string [] | string,
-                operator: ''
-              }
-              if (filter === 'genre' || filter === 'contentRating') {
-                const googlePlayStorValues: string[] = [];
-                const appStorValues: string[] = [];
-                let appStorInfo: any = {
-                  name: '',
-                  value: '' as String[] | string,
-                  operator: ''
-                };
-                let googlePlayStorInfo: any = {
-                  name: '',
-                  value: '' as String[] | string,
-                  operator: ''
-                };
-                tableFilters[filter]?.forEach((key: string, index: number, data: string[]) => {
-                  const filterValues = key.split('-');
-                  if (data?.length === 1) {
-                    singleFilter = {
-                      name:  filterValues[0],
-                      value: filterValues[1],
-                      operator: filterValues[2]
-                    }
-                  } else {
-                    if(filterValues[0].split('.')[0] === 'googlePlayStoreInfo') {
-                      googlePlayStorValues.push(filterValues[1]);
-                      googlePlayStorInfo = {
-                        name: filterValues[0],
-                        value: googlePlayStorValues,
-                        operator: filterValues[2]
-                      }
-                    }
-                    if(filterValues[0].split('.')[0] === 'appStoreInfo') {
-                      appStorValues.push(filterValues[1]);
-                      appStorInfo = {
-                        name: filterValues[0],
-                        value: appStorValues,
-                        operator: filterValues[2]
-                      }
-                    }
-                  }
-                })
-                if (appStorInfo.value) {
-                  updatedFilters.push(appStorInfo);
-                }
-                if (googlePlayStorInfo.value) {
-                  updatedFilters.push(googlePlayStorInfo);
-                }
-              } else {
-                tableFilters[filter]?.forEach((key: string) => {
-                  const filterValues = key.split('?');
-                  console.log('filtervalues => ', filterValues);
-                  updatedFilters.push({
-                    name:  filterValues[0],
-                    value: filterValues[1],
-                    operator: filterValues[2]
-                  });
-                })
-              }
-            }
-          });
-        }
+        const updatedFilters: IFilter[] = useFilters(tableFilters);
         setRequestBody({...requestBody, filters: updatedFilters});
     }
   }
